feat: sprite component support custom geometry
